+++ lib/premailex/html_parser/meeseeks.ex
@@ -27,8 +27,8 @@ if Code.ensure_loaded?(Meeseeks) do
       end
     end
 
-    # Meeseeks wraps fragments in an <html> element, so we unwrap it if the
-    # input was a fragment.
+    # Meeseeks wraps fragments in an <html> with <head> and <body> tags, so we
+    # must unwrap it if the input was a fragment.
     defp unwrap_fragment(tree, html) do
       case Regex.match?(~r/<html/i, html) do
         true ->
@@ -37,9 +37,9 @@ if Code.ensure_loaded?(Meeseeks) do
         false ->
           # This may break if Meeseeks changes how it wraps fragments. If so,
           # move this into a function that handles different Meeseeks versions.
-          [{"html", [], [{"head", [], []}, {"body", [], fragment}]}] = tree
+          [{"html", [], [{"head", [], head}, {"body", [], body}]}] = tree
 
-          fragment
+          head ++ body
       end
     end